Fix scaling bug first manifesting in PySide 6.7+
Previously the 'metric' method of OutlinePaintDevice would return the same
values for PdmDevicePixelRatioScaled and PdmDevicePixelRatio despite the fact
that the (non-overridden) scale factor returned by devicePixelRatioFScale was
not 1.

In principle we should ideally override devicePixelRatioFScale since we can
natively handle real values but this is an inlined method in the C++ and is not
overridable from Python. As a result, we are forced to stick to the default
value (which happens to be 1/0x10000). Since 1/0x10000 is an exact power of two
scaling a floating point number by this value only effects the exponent and,
assuming we're not too close to the precision limits (which we shouldn't be)
should have no effect on numerical accuracy.

The reason this bug only manifests in PySide 6.7+ is that Qt had previously
clamped the device scaling factor at 1 -- a value which makes sense for most
screens -- but a recent change [0] removed this clamp to support arbitrarily
scaled environments. As a result, the less-than-1 default scaling factor
was suddenly applied, revealing this bug.

[0] https://github.com/qt/qtbase/commit/9e740c7266fb72c3e9dfcfde9912510180b600c2

Many thanks to @gabrieldavison on GitHub for identifying this bug!

Fixes #4